fix: 회원탈퇴 실패 오류 수정 (s3이미지 경로 삭제 중 null을 참조해서 생긴오류)

+++ src/main/java/com/wordwave/myvocabook/MyVocaBookService.java
@@ -148,8 +148,10 @@ public class MyVocaBookService {
 	public void deleteAllImageURL(SiteUser user) throws Exception {
 		List<MyVocaBook> myVocaBooks = myVocaBookRepository.findAllByUser(user);
 		for(MyVocaBook mvb : myVocaBooks) {
-			String keyName = UrlParser.getKeyFromUrl(mvb.getImageURL());
-			s3Service.delete(keyName);
+			if(mvb.getImageURL() != null) {
+				String keyName = UrlParser.getKeyFromUrl(mvb.getImageURL());
+				s3Service.delete(keyName);
+			}
 		}
 	}
